Polished the navbar and added back scroll-snapping
did some black magic to fix the interaction between the fixed navbar and the css scroll snapping causing a blackhole in space time. (Added a div with some adequate padding to provide enough room between the start of the first section and the navbar. Hard to explain (I don't even know precisely what I did there).
Then removed the call-to-action swipe up animation on index.html because it overlapped the toggled navbar, It wouldn't be fixed by adding a z-index to both.